Address #43 Rare bug in PDF417 barcode generator
This implements suggested fix for bug reported to SourceForge project.

If a value to be enceded as PDF417 contains a run of 45 digits (0-9) and invalid barcode is generated. The issue is rare in that it happens only when there are exactly 45 consequtive digits. For example, error happens with 45 digits but not with 44 or 46.

Original bug report:
https://sourceforge.net/p/barcode4j/bugs/43/